requirements.txt
Flask>=2.3
flask-cors>=4.0
gunicorn>=21.2
python-dotenv>=1.0
requests>=2.31
pymongo>=4.6
groq>=0.9.0
httpx>=0.27
pdfplumber
python-docx
reportlab
speechrecognition
pyttsx3
sounddevice
vosk
sounddevice
numpy
feedparser
